+++ src/lib.rs
@@ -102,6 +102,8 @@ extern crate anyhow;
 
 pub mod conf;
 mod run;
+
+#[macro_use]
 mod util;
 
 use flood_tide::HelpVersion;
@@ -174,5 +176,5 @@ pub fn execute_env(
             return Err(anyhow!("{}\n{}", errs, TRY_HELP_MSG));
         }
     };
-    run::run(sioe, &conf, &env)
+    run::run(sioe, &conf, env)
 }