+++ tests/test_s.rs
@@ -3,41 +3,45 @@ macro_rules! help_msg {
         concat!(
             version_msg!(),
             "\n",
-            "Usage:\n",
-            "  aki-gsub [options]\n",
-            "\n",
-            "substitude text command, replace via regex.\n",
-            "\n",
-            "Options:\n",
-            "      --color <when>    use markers to highlight the matching strings\n",
-            "  -e, --exp <exp>       regular expression\n",
-            "  -f, --format <fmt>    replace format\n",
-            "  -n, --quiet           no output unmach lines\n",
-            "\n",
-            "  -H, --help        display this help and exit\n",
-            "  -V, --version     display version information and exit\n",
-            "  -X <x-options>    x options. try -X help\n",
-            "\n",
-            "Option Parameters:\n",
-            "  <when>    'always', 'never', or 'auto'\n",
-            "  <exp>     regular expression can has capture groups\n",
-            "  <fmt>     format can has capture group: $0, $1, $2, ...\n",
-            "\n",
-            "Environments:\n",
-            "  AKI_GSUB_COLOR_SEQ_ST     color start sequence specified by ansi\n",
-            "  AKI_GSUB_COLOR_SEQ_ED     color end sequence specified by ansi\n",
-            "\n",
-            "Examples:\n",
-            "  Leaving one character between 'a' and 'c', converts 'a' and 'c'\n",
-            "  on both sides to '*':\n",
-            "    echo \"abcabca\" | aki-gsub -e \"a(.)c\" -f \"*\\$1*\"\n",
-            "  result output:\n",
-            "    *b**b*a\n",
-            "\n",
-            "  Converts 'a' to '*' and 'c' to '@':\n",
-            "    echo \"abcabca\" | aki-gsub -e \"a\" -f \"*\" -e \"c\" -f \"@\"\n",
-            "  result output:\n",
-            "    *b@*b@*\n",
+            indoc::indoc!(
+                r#"
+            Usage:
+              aki-gsub [options]
+
+            substitude text command, replace via regex.
+
+            Options:
+                  --color <when>    use markers to highlight the matching strings
+              -e, --exp <exp>       regular expression
+              -f, --format <fmt>    replace format
+              -n, --quiet           no output unmach lines
+
+              -H, --help        display this help and exit
+              -V, --version     display version information and exit
+              -X <x-options>    x options. try -X help
+
+            Option Parameters:
+              <when>    'always', 'never', or 'auto'
+              <exp>     regular expression can has capture groups
+              <fmt>     format can has capture group: $0, $1, $2, ...
+
+            Environments:
+              AKI_GSUB_COLOR_SEQ_ST     color start sequence specified by ansi
+              AKI_GSUB_COLOR_SEQ_ED     color end sequence specified by ansi
+
+            Examples:
+              Leaving one character between 'a' and 'c', converts 'a' and 'c'
+              on both sides to '*':
+                echo "abcabca" | aki-gsub -e "a(.)c" -f "*\$1*"
+              result output:
+                *b**b*a
+
+              Converts 'a' to '*' and 'c' to '@':
+                echo "abcabca" | aki-gsub -e "a" -f "*" -e "c" -f "@"
+              result output:
+                *b@*b@*
+            "#
+            ),
             "\n",
         )
     };