+++ CHANGELOG.md
@@ -2,6 +2,15 @@ aki-gsub TBD
 ===
 Unreleased changes. Release notes have not yet been written.
 
+0.1.32 (2021-09-11)
+=====
+
+* pass cargo clippy
+* update depends: anyhow(1.0.43), flood-tide-gen(0.1.14), flood-tide(0.2.3), memx-cdy(0.1.6), runnel(0.3.8)
+* rewite TARGET_EXE_PATH with `env!(concat!("CARGO_BIN_EXE_", env!("CARGO_PKG_NAME")))`
+* update depends: exec-target(0.2.3)
+* add depends: indoc(1.0.3)
+
 0.1.31 (2021-06-24)
 =====